New GroceryCounter program and Tests
A GroceryCounter program was added that adds numbers up like a grocery counter. Tests were implemented to ensure each digit of the counter worked as expected. Each digit of the counter is pressed as many times as the user inputs and resets every time it reaches 10 and adds one to the next digit. Ones the tens digit reaches 10, one is added to the overflow counter and the tens digit resets to 0.